+++ src/widgets/mod.rs
@@ -1,0 +1,7 @@
+/// Container for dockable tabs.
+pub mod dock_area;
+
+pub(crate) mod popup;
+
+/// Trait for tab-viewing types.
+pub mod tab_viewer;
